refactor: hooks切り離し
